// src/types/store/adminAndTeacher/activity.d.ts
import { IActivity } from '../../../model/activity';

export type ActivityState = {
  adminActivityDoubleList: any[];
  activityList: any[];
  adminResourcesList: any[];
  classesList: any[];
  studentsList: any[];
  teachersList: any[];
  adminRoomsList: any[];
  activityItemSelected: any;
  activityAttendance: any[];
  activityAttendanceChanged: any;
  activityListChanged: any;
};