// src/model/activity.ts
export interface IActivity {
  id: string;
  title: string;
  type: string;
  showInSchedule: string;
  reserveParticipants: string;
  start: string;
  end: string;
  class: any[];
  teachers: any[];
  rooms: any[];
  resources: any[];
  status: string;
  account: string;
  note: string;
  students: any[];
  steachers: string;
  srooms: string;
  sresources: string;
  sclass: string;
  roomsresources: any[];
}